refactor: simplify lexer
